third_party/WebKit/LayoutTests/usb/usb-connection-event.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script src="resources/webusb-test.js"></script>
<script>
'use strict';

usb_test(() => {
  return getFakeDevice().then(({ device }) => {
    let evt = new USBConnectionEvent('connect', { device: device });
    assert_equals(evt.type, 'connect');
    assert_equals(evt.device, device);
  });
}, 'Can construct a USBConnectionEvent with a device');

test(t => {
  try {
    new USBConnectionEvent('connect', { device: null });
    assert_unreached('expected TypeError');
  } catch (err) {
    assert_true(err instanceof TypeError);
  }

  try {
    new USBConnectionEvent('connect', {})
    assert_unreached('expected TypeError');
  } catch (err) {
    assert_true(err instanceof TypeError);
  }
}, 'Cannot construct a USBConnectionEvent without a device');
</script>
